Atualizações afapSaude

+++ src/components/afapSaude/consultas.tsx
@@ -6,15 +6,20 @@ import { Badge, Button, Table } from "flowbite-react";
 import { ChangeEvent, useCallback, useContext, useRef, useState } from "react";
 import { ModalConsulta } from "./components/modalNovaConsulta";
 import { toast } from "react-toastify";
-import { HiDocument, HiPencil } from "react-icons/hi2";
+import { HiDocument, HiMiniArrowDownOnSquare, HiPencil, HiPrinter, HiTrash } from "react-icons/hi2";
 import { ModalDeletarExame } from "./components/modalDeletarExame";
-import {  HiFilter } from "react-icons/hi";
+import {  HiAdjustments, HiDocumentAdd, HiFilter } from "react-icons/hi";
 import { ModalFiltroConsultas } from "./components/modalFiltro";
 import FichaConsulta from "@/Documents/afapSaude/fichaConsulta";
 import { useReactToPrint } from "react-to-print";
 import { MdDelete } from "react-icons/md";
 import { BiMoneyWithdraw } from "react-icons/bi";
 import { AuthContext } from "@/contexts/AuthContext";
+import { GiReturnArrow } from "react-icons/gi";
+import { FaWhatsapp } from "react-icons/fa";
+import handleWhatsAppClick from "@/utils/openWhats";
+import pageStyle from "@/utils/pageStyle";
+import { ReciboMensalidade } from "@/Documents/mensalidade/Recibo";
 
 
 
@@ -28,6 +33,8 @@ interface DataProps {
   loading: boolean
 }
 export default function Consultas({ medicos, consultas, exames, setConsultas, buscarConsultas, loading }: DataProps) {
+
+  const valorInicial ={celular:'',cpf:'',data:new Date(),espec:'',exames:[],id_consulta:null,id_med:null,nome:'',tipoDesc:'',vl_consulta:0,vl_desc:0,vl_final:0}
   const [openModal, setOpenModal] = useState(false);
   const [data, setData] = useState<Partial<ConsultaProps>>()
   const {usuario} = useContext(AuthContext)
@@ -47,48 +54,53 @@ export default function Consultas({ medicos, consultas, exames, setConsultas, bu
   })
 
 const currentPage = useRef<FichaConsulta>(null)
+const currentRecibo = useRef<ReciboMensalidade>(null)
 
 
 
 const imprimirFicha = useCallback(useReactToPrint({
-  pageStyle: `
-  @page {
-      margin: 1rem;
-  }
-  @media print {
-      body {
-          -webkit-print-color-adjust: exact;
-      }
-      @page {
-          size: auto;
-          margin: 1rem;
-      }
-      @page {
-          @top-center {
-              content: none;
-          }
-          @bottom-center {
-              content: none;
-          }
-      }
-  }
-`,
+  pageStyle: pageStyle,
   content: () => currentPage.current,
 }), []);
 
-const handleReceberConsulta = useCallback(async (item: Partial<ConsultaProps>)=>{
+
+
+const imprimirRecibo = useCallback(useReactToPrint({
+  pageStyle: pageStyle,
+  content: () => currentRecibo.current,
+}), []);
+
+
+
+
+
+
+
+const handleReceberConsulta = useCallback(async ()=>{
+
+
+  if(!data?.id_consulta){
+    toast.warning('Selecione uma consulta')
+    return
+  }
+
+  if(data?.status === 'RECEBIDO'){
+    toast.warning('Consulta ja foi recebida!')
+    return
+  }
 
 try {
   const dataAtual = new Date()
-  dataAtual.setHours(dataAtual.getHours() - dataAtual.getTimezoneOffset() / 60)
+  dataAtual.setTime(dataAtual.getTime() - dataAtual.getTimezoneOffset() * 60 * 1000)
+
   const response = await toast.promise(
     api.put('/afapSaude/receberConsulta',{
-      id_consulta: item?.id_consulta,
+      id_consulta: data?.id_consulta,
     id_usuario:usuario?.id,
-    datalancUTC:dataAtual,
+    datalancUTC:dataAtual.toISOString(),
     descricao:"CONSULTA",
-    historico:`CONSULT.${item?.id_consulta}-${item?.nome}-${item?.espec}`,
-    valor:item?.vl_final,
+    historico:`CONSULT.${data?.id_consulta}-${data?.nome}-${data?.espec}`,
+    valor:data?.vl_final,
     usuario:usuario?.nome,
     }),
     {
@@ -98,6 +110,7 @@ try {
     }
   )
   buscarConsultas({startDate:new Date(),endDate:new Date()})
+  setData(valorInicial)
 } catch (error) {
   console.log(error)
 }
@@ -106,6 +119,11 @@ try {
 
 
 const handleDeletar = useCallback(async () => {
+
+  if(!data?.id_consulta){
+    toast.warning('Selecione uma consulta')
+    return
+  }
   try {
     const response = await toast.promise(
       api.delete(`/afapSaude/consultas/deletarCadastro/${data?.id_consulta}`),
@@ -119,7 +137,11 @@ const handleDeletar = useCallback(async () => {
  
   } catch (error) {
     console.log(error);
+
   }
+  setModalDeletar(false);
+  setData(valorInicial);
+
 }, [consultas, data, setConsultas]);
 
 
@@ -144,6 +166,9 @@ const handleDeletar = useCallback(async () => {
     const item = exames.find(atual => atual.id_exame === Number(event.target.value))
 
 
+    
+
+
     item?.id_exame && setDataExam({
       data: new Date(),
       id_exame: item?.id_exame,
@@ -162,15 +187,49 @@ const handleDeletar = useCallback(async () => {
   return (
     <div className="flex flex-col p-2 gap-2">
       <div className="ml-auto inline-flex gap-4">
-        <Button theme={{ color: { light: "border border-gray-300 bg-white text-gray-900  enabled:hover:bg-gray-100 " } }} color={'light'} size={'sm'} onClick={() => setModalFiltro(true)}>  <HiFilter className="mr-2 h-5 w-5" /> Filtro</Button>
-        <Button size={'sm'} onClick={() =>{setData({celular:'',cpf:'',data:new Date(),espec:'',exames:[],id_consulta:null,id_med:null,nome:'',tipoDesc:'',vl_consulta:0,vl_desc:0,vl_final:0}), setOpenModal(true)}}>Nova Consulta</Button>
+      
+       
+        <Button.Group >
+      
+      <Button onClick={() =>{setData(valorInicial), setOpenModal(true)}} size={'sm'} color="gray">
+        <HiDocumentAdd className="mr-2 h-4 w-4" />
+        Adicionar
+      </Button>
+      <Button onClick={() =>  setOpenModal(true)} size={'sm'} color="gray">
+        <HiPencil className="mr-2 h-4 w-4" />
+        Editar
+      </Button>
+      <Button onClick={() => imprimirFicha()} size={'sm'} color="gray">
+        <HiDocument className="mr-2 h-4 w-4" />
+         Prontuário
+      </Button>
+
+      <Button onClick={imprimirRecibo} size={'sm'} color="gray">
+        <BiMoneyWithdraw className="mr-2 h-4 w-4" />
+         Recibo
+      </Button>
+      <Button onClick={() => handleReceberConsulta()} size={'sm'} color="gray">
+        <HiMiniArrowDownOnSquare className="mr-2 h-4 w-4" />
+        Receber
+      </Button>
+      <Button color="gray" type="button"  ><GiReturnArrow className="mr-2 h-4 w-4"/> Estornar</Button>
+      <Button onClick={()=>handleWhatsAppClick(data?.celular)} size={'sm'} color="gray">
+      <FaWhatsapp className="mr-2 h-4 w-4" />
+        Abrir Conversa
+      </Button>
+      <Button onClick={()=>setModalDeletar(true)} size={'sm'} color="gray">
+        <MdDelete className="mr-2 h-4 w-4" />
+        Excluir
+      </Button>
+    </Button.Group>
+    <Button theme={{ color: { light: "border border-gray-300 bg-white text-gray-900  enabled:hover:bg-gray-100 " } }} color={'light'} size={'sm'} onClick={() => setModalFiltro(true)}>  <HiFilter className="mr-2 h-4 w-4" /> Filtro</Button>
       </div>
 
 
 
 
       <div className="overflow-x-auto h-[calc(100vh-160px)]">
-        <Table hoverable theme={{ body: { cell: { base: "px-6 text-black py-2 group-first/body:group-first/row:first:rounded-tl-lg group-first/body:group-first/row:last:rounded-tr-lg group-last/body:group-last/row:first:rounded-bl-lg group-last/body:group-last/row:last:rounded-br-lg text-xs" } } }}  >
+        <Table  theme={{ body: { cell: { base: "px-6 text-black py-2 group-first/body:group-first/row:first:rounded-tl-lg group-first/body:group-first/row:last:rounded-tr-lg group-last/body:group-last/row:first:rounded-bl-lg group-last/body:group-last/row:last:rounded-br-lg text-xs" } } }}  >
 
           <Table.Head>
             <Table.HeadCell>Nome</Table.HeadCell>
@@ -179,14 +238,11 @@ const handleDeletar = useCallback(async () => {
             <Table.HeadCell>Tipo Desc.</Table.HeadCell>
             <Table.HeadCell>Valor Desc.</Table.HeadCell>
             <Table.HeadCell>Valor Final</Table.HeadCell>
-            <Table.HeadCell>Status</Table.HeadCell>
-            <Table.HeadCell>
-              <span className="sr-only">Edit</span>
-            </Table.HeadCell>
+            <Table.HeadCell>Status</Table.HeadCell>    
           </Table.Head>
           <Table.Body className="divide-y">
             {consultas.map((item, index) => (
-              <Table.Row key={item.id_consulta} className="bg-white dark:border-gray-700 dark:bg-gray-800">
+              <Table.Row onClick={() => setData(item)} key={item.id_consulta} className={`bg-white hover:cursor-pointer ${data?.id_consulta === item.id_consulta ? 'bg-gray-400' : ''} `}>
                 <Table.Cell className="whitespace-nowrap font-medium text-gray-900 dark:text-white">
                   {item.nome}
                 </Table.Cell>
@@ -197,22 +253,9 @@ const handleDeletar = useCallback(async () => {
                 <Table.Cell>{Number(item.vl_final).toLocaleString('pt-BR', { style: 'currency', currency: 'BRL' })}</Table.Cell>
                 <Table.Cell>
                 
-              <Badge color={item.status==='PENDENTE'?'yellow':item.status==='CANCELADO'?'red':'green'}>{item.status}</Badge>
+              <Badge className="justify-center" color={item.status==='PENDENTE'?'yellow':item.status==='CANCELADO'?'red':'green'}>{item.status}</Badge>
                 </Table.Cell>
-                <Table.Cell className="space-x-4">
-                  <button onClick={() => { setData({ ...item }), setOpenModal(true) }} className="font-medium text-gray-500 hover:text-cyan-600">
-                    <HiPencil size={16} />
-                  </button>
-                  <button onClick={() => { setModalDeletar(true), setData({ ...item }) }} className="font-medium text-gray-500 hover:text-red-600 ">
-                    <MdDelete size={18} />
-                  </button>
-                  <button onClick={() => imprimirFicha()} className="font-medium text-gray-500 hover:text-blue-600 ">
-                    <HiDocument size={16} />
-                  </button>
-                  <button onClick={() => handleReceberConsulta(item)} className="font-medium text-gray-500 hover:text-blue-600 ">
-                    <BiMoneyWithdraw size={18} />
-                  </button>
-                </Table.Cell>
+            
 
               </Table.Row>
             ))}
@@ -228,7 +271,35 @@ const handleDeletar = useCallback(async () => {
 
 
       <div className="hidden" style={{display:'none'}}>
-        <FichaConsulta bairro="" cidade="" contrato={0} cpf="" logradouro="" material="" nome="" rg="" telefone="" uf="" ref={currentPage}/>
+        <FichaConsulta 
+        bairro={data?.bairro??''}
+         cidade={data?.cidade??''} 
+          cpf={data?.cpf??''}
+          endereco={data?.endereco??''} 
+          nascimento={data?.nascimento??undefined}
+          responsavel={data?.responsavel??''}
+           nome={data?.nome??''}
+            rg=""
+             celular={data?.celular??''}
+             parentesco={data?.grau_parentesco??''}
+              
+               ref={currentPage}/>
+
+
+               <ReciboMensalidade
+                associado={data?.nome??''}
+                contrato={data?.id_consulta??null}
+                data_pgto={data?.dt_pgto??null}
+                n_doc=""
+                referencia=""
+                valor={data?.vl_final??0}
+                vencimento={new Date()}
+                ref={currentRecibo}
+                referente={`Consulta Médica - ${data?.espec}`}
+
+               
+               />
+              
       </div>
     </div>
   );
